--- webapp/views/App/views/Home/Dashboard/RecordsSummary/Chart/index.js
@@ -1,1 +1,0 @@
-export { default } from './Chart'
