+++ webapp/views/App/views/Home/Dashboard/RecordsSummaryPeriodSelector/RecordsSummaryPeriodSelector.js
