+++ webapp/views/App/views/Home/Dashboard/RecordsSummaryPeriodSelector/index.js
