+++ webapp/views/App/views/Home/Dashboard/RecordsSummaryPeriodSelector/store/useTimeRanges.js
