+++ webapp/views/App/views/Home/Dashboard/RecordsSummaryPeriodSelector/store/utils/getFromDate.js
