+++ webapp/views/App/views/Home/Dashboard/TotalRecordsSummaryChart/TotalRecordsSummaryChart.scss
