+++ webapp/views/App/views/Home/Dashboard/TotalRecordsSummaryChart/components/dataPath.js
