+++ webapp/views/App/views/Home/Dashboard/TotalRecordsSummaryChart/components/xAxis.js
