+++ webapp/views/App/views/Home/Dashboard/TotalRecordsSummaryChart/index.js
@@ -1,0 +1,1 @@
+export { default } from './TotalRecordsSummaryChart'
